fix the Long to Int conversion
git-svn-id: file:///semplestb/final_repo@1723 02200ff9-b5b2-46f0-9171-221b09c08c7b

+++ SemplestServices/src/semplest/server/service/springjdbc/storedproc/UpdateRemainingBudgetInCycleSP.java
@@ -36,7 +36,7 @@ public class UpdateRemainingBudgetInCycleSP  extends StoredProcedure
 		}
 		else
 		{
-			return (Integer) results.get("ID");
+			return ((Long)results.get("ID")).intValue();
 		}
 	}
 }
